Version 1.0
Changed the name of the previous vimrc to basic_vimrc;
whithin it: functions were reformatted, useful stuff was added a bit
everywhere (mostly in search and functions).

extensions_vimrc and _vimrc were created: the file which will contain
all extensions' configurations and the real _vimrc file which should be
used.
Consequently the readme was updated with instructions.